feat: efficient internal state structure
 * Refactored entity state to nest related data in child objects
 * Updated all selectors to support new state structure
 * Refactored entity reducer to support new state structure
 * Decomposed monolithic reducer into multiple reducers for each set of actions

NAE-111
