+++ projects/ngrx-auto-entity/src/index.ts
@@ -132,7 +132,7 @@ export {
 /*
  * Reducer
  */
-export { autoEntityReducer, autoEntityMetaReducer, stateNameFromAction } from './lib/reducer/reducer';
+export { autoEntityReducer} from './lib/reducer/reducer';
 
 /*
  * Entity Service
@@ -186,3 +186,5 @@ export {
   UpsertManyEffect
 } from './lib/effects/effects-cud-discrete';
 export { NGRX_AUTO_ENTITY_APP_STORE } from './lib/effects/if-necessary-operator-utils';
+export { stateNameFromAction } from './lib/reducer/reduction.utils';
+export { autoEntityMetaReducer } from './lib/reducer/meta-reducer';
