+++ projects/ngrx-auto-entity/src/lib/decorators/entity-util.ts
@@ -1,5 +1,5 @@
 import { camelCase } from '../../util/case';
-import { pipe } from '../../util/func';
+import { compose } from '../../util/func';
 import { TNew } from '../actions/model-constructor';
 import { EntityComparer, IEntityOptions, IEntityTransformer } from './entity-options';
 import { ENTITY_OPTS_PROP } from './entity-tokens';
@@ -17,7 +17,7 @@ export const getEntityOptions = <TModel>(entityOrType: TNew<TModel> | TModel | T
     EMPTY_OBJECT) as IEntityOptions;
 
 export const entityOptions = <TModel>(entityOrType: TNew<TModel> | TModel | TModel[]): IEntityOptions =>
-  pipe(getEntity, ensureObject, getEntityOptions)(entityOrType);
+  compose(getEntity, ensureObject, getEntityOptions)(entityOrType);
 
 export const entityStateName = (modelName: string): string => camelCase(modelName);
