+++ projects/ngrx-auto-entity/src/lib/util/entity-state.ts
@@ -10,19 +10,34 @@ export interface IEntityDictionary<TModel> {
 }
 
 /**
- * Structure for how entities are stored, including useful computed properties
- * such as an array of their keys, status flags, timestamps, etc.
+ * Structure for how entity selections are stored
  */
-export interface IEntityState<TModel> {
-  entities: IEntityDictionary<TModel>;
-  ids: EntityIdentity[];
+export interface IEntitySelections {
   currentEntityKey?: EntityIdentity;
   currentEntitiesKeys?: EntityIdentity[];
+}
+
+/**
+ * Structure for how entity edits are stored
+ */
+export interface IEntityEdits<TModel> {
   editedEntity?: Partial<TModel>;
   isDirty?: boolean;
+}
+
+/**
+ * Structure for how entity paging info is stored
+ */
+export interface IEntityPaging {
   currentPage?: Page;
   currentRange?: Range;
   totalPageableCount?: number;
+}
+
+/**
+ * Structure for how entity tracking is stored
+ */
+export interface IEntityTracking {
   isLoading?: boolean;
   isSaving?: boolean;
   isDeleting?: boolean;
@@ -32,4 +47,17 @@ export interface IEntityState<TModel> {
   updatedAt?: number;
   replacedAt?: number;
   deletedAt?: number;
+}
+
+/**
+ * Structure for how entities are stored, including useful computed properties
+ * such as an array of their keys, status flags, timestamps, etc.
+ */
+export interface IEntityState<TModel> {
+  entities: IEntityDictionary<TModel>;
+  ids: EntityIdentity[];
+  selections?: IEntitySelections;
+  edits?: IEntityEdits<TModel>;
+  paging?: IEntityPaging;
+  tracking?: IEntityTracking;
 }
