+++ projects/ngrx-auto-entity/src/lib/util/selector-map.ts
@@ -20,22 +20,22 @@ export interface ISelectorMap<TParentState, TModel> {
   selectTotal: MemoizedSelector<object | TParentState, number>;
   selectHasEntities: MemoizedSelector<object | TParentState, boolean>;
   selectHasNoEntities: MemoizedSelector<object | TParentState, boolean>;
-  selectCurrentEntity: MemoizedSelector<object | TParentState, TModel | null>;
-  selectCurrentEntityKey: MemoizedSelector<object | TParentState, EntityIdentity | null>;
+  selectCurrentEntity: MemoizedSelector<object | TParentState, TModel | undefined>;
+  selectCurrentEntityKey: MemoizedSelector<object | TParentState, EntityIdentity | undefined>;
   selectCurrentEntities: MemoizedSelector<object | TParentState, TModel[]>;
   selectCurrentEntitiesKeys: MemoizedSelector<object | TParentState, EntityIdentity[]>;
-  selectEditedEntity: MemoizedSelector<object | TParentState, Partial<TModel> | null>;
+  selectEditedEntity: MemoizedSelector<object | TParentState, Partial<TModel> | undefined>;
   selectIsDirty: MemoizedSelector<object | TParentState, boolean>;
-  selectCurrentPage: MemoizedSelector<object | TParentState, Page | null>;
-  selectCurrentRange: MemoizedSelector<object | TParentState, Range | null>;
+  selectCurrentPage: MemoizedSelector<object | TParentState, Page | undefined>;
+  selectCurrentRange: MemoizedSelector<object | TParentState, Range | undefined>;
   selectTotalPageable: MemoizedSelector<object | TParentState, number>;
   selectIsLoading: MemoizedSelector<object | TParentState, boolean>;
   selectIsSaving: MemoizedSelector<object | TParentState, boolean>;
   selectIsDeleting: MemoizedSelector<object | TParentState, boolean>;
-  selectLoadedAt: MemoizedSelector<object | TParentState, Date | null>;
-  selectSavedAt: MemoizedSelector<object | TParentState, Date | null>;
-  selectCreatedAt: MemoizedSelector<object | TParentState, Date | null>;
-  selectUpdatedAt: MemoizedSelector<object | TParentState, Date | null>;
-  selectReplacedAt: MemoizedSelector<object | TParentState, Date | null>;
-  selectDeletedAt: MemoizedSelector<object | TParentState, Date | null>;
+  selectLoadedAt: MemoizedSelector<object | TParentState, Date | undefined>;
+  selectSavedAt: MemoizedSelector<object | TParentState, Date | undefined>;
+  selectCreatedAt: MemoizedSelector<object | TParentState, Date | undefined>;
+  selectUpdatedAt: MemoizedSelector<object | TParentState, Date | undefined>;
+  selectReplacedAt: MemoizedSelector<object | TParentState, Date | undefined>;
+  selectDeletedAt: MemoizedSelector<object | TParentState, Date | undefined>;
 }
